Specify exact version of conan to use in CI
This is one step towards more consistent builds by making the docker
containers independent from when they were built. Previously always the latest
version of conan was installed. Now it's fixed to a certain version.

This only affects the windows docker containers. The Linux docker
containers need to be revamped at a later point in time.